refactor: reduce complexity and eliminate code duplication in CharacterValidationForm
- Extract shared constants to constants.ts (CHARACTER_TYPE_OPTIONS, CHARACTER_RACE_OPTIONS, SIZE_OPTIONS, CHARACTER_CLASS_OPTIONS, ABILITY_SCORES, DEFAULT_CHARACTER_VALUES)
- Create shared utilities in utils.ts (getHitDieForClass, getAbilityModifier, formatModifier, transformToPreviewFormat)
- Update all form sections to use shared constants and utilities
- Reduce component complexity by eliminating duplicated code patterns
- Maintain full functionality and test coverage
